src/requirements.txt
Flask == 0.10.1
WTForms == 2.1
GitPython == 1.0.2
Requests == 2.9.1
PyYAML == 3.11
